Implement BER parsing of flights
Now we can parse the binary into a vector of flights/samples. The
parsing is quite hard coded and the BER encoding seems slightly strange
with some extra (4, 8) bytes in the start of each tuple. But it gives
correct samples.

So it can be made to look nice and probably faster in the future, but
now it works.