Implement hierarchical clustering
Now there is a function which takes in the matrix on vector form as well
as the number of "highest singles" we want, and returns that many of the
highest singles from the resulting dendrogram.

The reason it did not for example return airports or something more
descriptive is because this way it can be kept more generic.

The for loop is not pretty.